moved the contents of tgc_tess.c and tgc_tnurb.c into tgc.c
this works around an individuality of the MS Visual C/C++ linker, which
removes the content of seemingly unused .obj files from the resulting
library; this way, the rt_tgc_tess() and rt_tgc_tnurb() function symbols
were removed from the librt-static.lib on MS Windows
